pyproject.toml
[project]
name = "codex-a2a-server"
version = "0.1.0"
description = "A2A wrapper service for codex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk==0.3.22",
    "fastapi>=0.110",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sse-starlette>=2.1",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
codex-a2a-server = "codex_a2a_server.app:main"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.uv]
package = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:'HTTP_413_REQUEST_ENTITY_TOO_LARGE' is deprecated. Use 'HTTP_413_CONTENT_TOO_LARGE' instead.:DeprecationWarning:a2a\\.server\\.apps\\.jsonrpc\\.fastapi_app",
]
